# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "sakalWeb Information System v5 RESTful API documentation",
        "title": "swis-api v5",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://savla.dev/swapi",
            "email": "info@savla.dev"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/savla-dev/swis-api/blob/master/LICENSE"
        },
        "version": "5.2.4"
    },
    "host": "swapi.savla.su:8049",
    "basePath": "/",
    "paths": {
...
        "/business": {
            "get": {
                "description": "get business complete list",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "business"
                ],
                "summary": "Get all businesses",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/business.BusinessArray"
                        }
                    }
                }
            },
            "post": {
                "description": "add new business",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "business"
                ],
                "summary": "Add new business to the array",
                "parameters": [
                    {
                        "description": "query params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/business.Business"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/business.Business"
                        }
                    }
                }
            }
        },
...
        "/links": {
            "get": {
                "description": "get links complete list",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "links"
                ],
                "summary": "Get all links",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/links.Link"
                        }
                    }
                }
            },
            "post": {
                "description": "add new link to links array",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "links"
                ],
                "summary": "Add new link to links",
                "parameters": [
                    {
                        "description": "query params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/links.Link"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/links.Link"
                        }
                    }
                }
            }
        },
...
        "/links/{hash}": {
            "get": {
                "description": "get link by its :hash param",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "links"
                ],
                "summary": "Get link by :hash",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/links.Link"
                        }
                    }
                }
            },
            "put": {
                "description": "update link by its Hash",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "links"
                ],
                "summary": "Update link by its Hash",
                "parameters": [
                    {
                        "description": "query params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/links.Link"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/links.Link"
                        }
                    }
                }
            },
            "delete": {
                "description": "delete link by its Hash",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "links"
                ],
                "summary": "Delete link by its Hash",
                "parameters": [
                    {
                        "type": "string",
                        "description": "link Hash",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/links.Link"
                        }
                    }
                }
            }
        },
        "/links/{hash}/active": {
            "put": {
                "description": "toggle active boolean for {hash}",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "links"
                ],
                "summary": "Toggle active boolean for {hash}",
                "parameters": [
                    {
                        "type": "string",
                        "description": "hash",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/links.Link"
                        }
                    }
                }
            }
        },
...
        "/projects": {
            "post": {
                "description": "add new project to projects list",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "projects"
                ],
                "summary": "Add new project",
                "parameters": [
                    {
                        "description": "query params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/projects.Project"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/projects.Project"
                        }
                    }
                }
            }
        },
...
        "/roles": {
            "get": {
                "description": "get roules complete list",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roles"
                ],
                "summary": "Get all roles",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/roles.Roles"
                        }
                    }
                }
            },
            "post": {
                "description": "add new role to roles array",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roles"
                ],
                "summary": "Add new role to roles array",
                "parameters": [
                    {
                        "description": "query params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/roles.Role"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/roles.Role"
                        }
                    }
                }
            }
        },
...
        "/users": {
            "get": {
                "description": "get users complete list",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get all users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/users.Users"
                        }
                    }
                }
            },
            "post": {
                "description": "add new user to users array",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Add new user to users array",
                "parameters": [
                    {
                        "description": "query params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/users.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/users.User"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "business.Business": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "contact": {
                    "description": "Array of contacts of different type.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/business.Contact"
                    }
                },
                "ico": {
                    "description": "Czech Company ICO/ID number.",
                    "type": "string"
                },
                "id": {
                    "description": "Busines unique identifier.",
                    "type": "string"
                },
                "name_label": {
                    "description": "Company's business name.",
                    "type": "string"
                },
                "role": {
                    "description": "Business role to such organization (e.g. partner, owner)",
                    "type": "string"
                },
                "username": {
                    "description": "User's name linked to such business.",
                    "type": "string"
                },
                "vat_id": {
                    "description": "Czech company DICO/VAT ID number/string.",
                    "type": "string"
                }
            }
        },
        "business.BusinessArray": {
            "type": "object",
            "properties": {
                "business": {
                    "description": "Array of business records.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/business.Business"
                    }
                }
            }
        },
        "business.Contact": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "Contact field contents.",
                    "type": "string"
                },
                "type": {
                    "description": "Type of contact field (e.g. e-mail address, street address, telephone number etc).",
                    "type": "string"
                }
            }
        },
...
        "links.Link": {
            "type": "object",
            "required": [
                "hash",
                "url"
            ],
            "properties": {
                "active": {
                    "description": "Link's activated status.",
                    "type": "boolean"
                },
                "description": {
                    "description": "Link's more verbose name/description.",
                    "type": "string"
                },
                "hash": {
                    "description": "Link's unique hash ID.",
                    "type": "string"
                },
                "name": {
                    "description": "Link's name/placeholder.",
                    "type": "string"
                },
                "url": {
                    "description": "Link's URL to link to.",
                    "type": "string"
                }
            }
        },
...
        "projects.Project": {
            "type": "object",
            "required": [
                "project_id"
            ],
            "properties": {
                "backuped": {
                    "description": "Projects backuped boolean.",
                    "type": "boolean"
                },
                "kanban_link": {
                    "description": "URL to kanboard/kanban project's page.",
                    "type": "string"
                },
                "project_deploy_target": {
                    "description": "Target internal node of deployment.",
                    "type": "string"
                },
                "project_desc": {
                    "description": "Brief project description.",
                    "type": "string"
                },
                "project_docs_link": {
                    "description": "URL to documentation page(s).",
                    "type": "string"
                },
                "project_id": {
                    "description": "Project's unique identificator.",
                    "type": "string"
                },
                "project_manager": {
                    "description": "Project manager's name/username.",
                    "type": "string"
                },
                "project_name": {
                    "description": "Project name.",
                    "type": "string"
                },
                "project_published": {
                    "description": "Published boolean.",
                    "type": "boolean"
                },
                "project_repo": {
                    "description": "Git repository link (not URL, without HTTP scheme).",
                    "type": "string"
                },
                "project_url": {
                    "description": "URL to base page of the project (project's URL).",
                    "type": "string"
                },
                "redmine_link": {
                    "description": "URL to redmine project overview.",
                    "type": "string"
                }
            }
        },
...
        "roles.Role": {
            "type": "object",
            "properties": {
                "active": {
                    "description": "Role status, by default it is inactive.",
                    "type": "boolean"
                },
                "description": {
                    "description": "Role description to make more sense when listing those.",
                    "type": "string"
                },
                "name": {
                    "description": "Role name is its unique description, acts like an ID too.",
                    "type": "string"
                }
            }
        },
        "roles.Roles": {
            "type": "object",
            "properties": {
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/roles.Role"
                    }
                }
            }
        },
...
        "users.User": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "active": {
                    "description": "Presence/Absence boolean. If false, one is not allowed to log-in (token is rejected),\nto interract with savla-dev infra in general (by default).",
                    "type": "boolean"
                },
                "country": {
                    "description": "Country of origin -- to help maintain global contacts.",
                    "type": "string"
                },
                "discord_username": {
                    "description": "Discord account/profile name.",
                    "type": "string"
                },
                "email_alias": {
                    "description": "Email alias in cloudflare email routing",
                    "type": "string"
                },
                "email_main": {
                    "description": "Email address main, personal",
                    "type": "string"
                },
                "full_name": {
                    "description": "Full Name of such user.",
                    "type": "string"
                },
                "gdpr_consent": {
                    "description": "Important GDPR consent boolean -- if false, user's details should be omitted!\nSEE more -- https://gdpr.eu/checklist/",
                    "type": "boolean"
                },
                "github_username": {
                    "description": "GitHub account/profile name (used for SSH public keys importing).",
                    "type": "string"
                },
                "gpg_keys": {
                    "description": "User's GPG public keys array.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "description": "ID not used anymore as indexing is used differently now (searching by Name, index respects array implicit property).\nID       \tstring \t\t`json:\"id\"`",
                    "type": "string"
                },
                "roles": {
                    "description": "User's given roles -- a role labels array.\nRoles []roles.Role `json:\"roles\"`",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "spotify_link": {
                    "description": "Spotify link to one's profile.",
                    "type": "string"
                },
                "ssh_keys": {
                    "description": "User's SSH public keys array.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "token_hmac": {
                    "description": "Unique token used for auth purposes, SHA?-HMAC.",
                    "type": "string"
                },
                "wireguard_vpn": {
                    "description": "All Wireguard config objects -- an array.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/users.Wireguard"
                    }
                }
            }
        },
        "users.Users": {
            "type": "object",
            "properties": {
                "users": {
                    "description": "An array of User objects",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/users.User"
                    }
                }
            }
        },
        "users.Wireguard": {
            "type": "object",
            "properties": {
                "allowed_ips": {
                    "description": "Allowed IP address(es) list on the side of server (vpn_gateway_server).",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "device_name": {
                    "description": "Unique device name (for such user).",
                    "type": "string"
                },
                "ip_address": {
                    "description": "User's private IP address.",
                    "type": "string"
                },
                "permission": {
                    "description": "Is the user given permission to dial a connection?",
                    "type": "boolean"
                },
                "private_key": {
                    "description": "Wireguard private key.\nTODO: should be encrypted?",
                    "type": "string"
                },
                "public_key": {
                    "description": "Wireguard public key.",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "apiKey": {
            "type": "apiKey",
            "name": "X-Auth-Token",
            "in": "header"
        }
    }
}